Increase ACCEL_MAX for PID But only at Low Speeds
This is a further attempt to solve the sluggish start issue on
TSS2 Toyotas. By allowing for a higher ACCEL_MAX for the PID
controller at low speeds < 6m/s.  I reverted baack ACCEL_MAX
to its original value for non-TSS2 vehicles and above 6m/s.

Rationale:
It looks to me, that part of the issue is that the PID maxs out
very quickly, but the max is not enough to cause the aEgo
to reach the desired accel quick enough.

Looking at a real-world example starting behind a lead that
accelerated away at ~1.5m/s2 (which is faster than stock
(1.2m/s2) would be capable of keeping up with):

0.0s = Brakes released, positive accel requested
0.7s = Initial movement of car
1.3s = PID Maxes out
3.8s = aEgo reaches desired accel